Documentation and got rid of some warnings.

+++ northstar-robomaster-project/src/control/flywheel/flywheel_subsystem.cpp
@@ -92,7 +92,7 @@ float FlywheelSubsystem::launchSpeedToFlywheelRpm(float launchSpeed) const
 {
     modm::interpolation::Linear<modm::Pair<float, float>> MPSToRPMInterpolator = {
         spinToRPMMap.at(desiredSpin).data(),
-        spinToRPMMap.at(desiredSpin).size()};
+        static_cast<uint8_t>(spinToRPMMap.at(desiredSpin).size())};
     return MPSToRPMInterpolator.interpolate(launchSpeed);
 }
